\name{phytotools-package}
\alias{phytotools-package}
\alias{phytotools}
\docType{package}
\title{
Phytoplankton Production Tools
}

\description{

Fits PE and RLC data to one of a four published PE models.
Simulates incident irradiance as a function of time and space.
Calculates phytoplankton production by transposing modeled PE or RLC data 
to a water column with a user-defined theoretical in-situ irradiance field.


}

\author{
Greg M. Silsbe
Sairah Y. Malkin

Maintainer: Greg Silsbe <gsilsbe@gmail.com>
}

\references{

Eilers, P.H.C. and Peeters, J.C.H. 1988 A model for the relationship between light intensity and the rate of photosynthesis in phytoplankton. \emph{Ecological Modeling}. \bold{42}, 199--215.

Fee, E.J. 1990. Computer programs for calculating in-situ phytoplankton photosynthesis. \emph{Canadian technical report for fisheries and aquatic sciences}. \bold{No. 1740}.

Hofierka, J., and Suri, M. 2002 The solar radiation model for Open source GIS: Implementation and applications. \emph{Proceedings of the source GIS-GRASS users conference 2002}. 

Jassby, A.D.and Platt, T. 1976 Mathematical formulation of the relationship between photosynthesis and light for phytoplankton. \emph{Limnology and Oceanography}. \bold{21}, 540--547.

Kirk, J.T.O., 2011 Light and photosynthesis in aquatic environments. Cambridge Press.

Platt, T., Gallegos, C.L. and Harrison, W.G. 1980 Photoinhibition and photosynthesis in natural assemblages of marine phytoplankton. \emph{Journal of Marine Research}. \bold{38}, 687--701.

Silsbe, G.M., and Kromkamp, J.C. 2012 Modeling the irradiance dependency of the quantum efficiency of photosynthesis. \emph{Limnology and Oceanography: Methods}. \bold{10}, 642--652. 

Soetaert, K. and Petzoldt, T. 2010 Inverse Modelling, Sensitivity and Monte Carlo Analysis in R Using Package FME. \emph{Journal of Statistical Software}. \bold{33}, 1?28.

Webb, W.L., Newton, M., and Starr, D. 1974 Carbon dioxide exchange of \emph{Alnus rubra}: A mathematical model. \emph{Oecologia}. \bold{17}, 281--291.

}